requirements.txt
streamlit
google-genai
requests
pdfplumber
PyMuPDF
pydantic
PyPDF2
